+++ src/main/java/com/flyn/eeg_receiver/listener/EEGPowerListener.java
@@ -1,9 +1,0 @@
-package com.flyn.eeg_receiver.listener;
-
-import com.flyn.eeg_receiver.event.EEGPowerEvent;
-
-public interface EEGPowerListener extends DeviceListener {
-
-    void onEEGPowerEvent(EEGPowerEvent event);
-
-}