+++ src/main/kotlin/com/flyn/eeg_receiver/listener/ExampleListener.kt
@@ -2,7 +2,8 @@ package com.flyn.eeg_receiver.listener
 
 import com.flyn.eeg_receiver.event.*
 
-object ExampleListener: PoorSignalListener, AttentionListener, MeditationListener, StrengthListener, RawWaveListener, EEGPowerListener {
+object ExampleListener: PoorSignalListener, AttentionListener, MeditationListener, StrengthListener, RawWaveListener,
+    SpectralListener {
 
     override fun onRawWaveEvent(event: RawWaveEvent) {
         println("Receive Raw Wave Event, value: ${event.value}")
@@ -16,7 +17,7 @@ object ExampleListener: PoorSignalListener, AttentionListener, MeditationListene
         println("Receive Attention Event, value: ${event.value}")
     }
 
-    override fun onEEGPowerEvent(event: EEGPowerEvent) {
+    override fun onSpectralEvent(event: SpectralEvent) {
         print("Receive EEG Event,")
         print("low-alpha: ${event.lowAlpha},")
         print("high-alpha: ${event.highAlpha},")